Revert changes in b3e7d1cf1fe1d3506f7ebb6cfd9312eab274b830
The changes are not a good idea at least at current.

Simply adding the PLS_PPM_SIZE will cause non-expected or error behaviors.

The reason is as follows.

1) D-thinker memory system should have a consistent view of the memory
space layout as the Puc programs. The macro here only reflects what the Puc
program will consider the memory layout but not what the D-thinker memory considers.
Non-consistency in the views of memory layout will cause non-expected behaviors.

2) D-thinker's memory should not be modified after it is set. If it is changed,
what if some non-persistent memory turns to persistent memory? Where is the data?
what if some non-replicated memory range turns to replicated memory range? Where is
the replicated pages? The metadata structures will need to be changed? How to change it?
How long will it take? and lots more problems.

For now, the added macro does not work as expected. Hence, I revert the changes.

The reverting is as a commit (this one). So when we want to add them back,
we can re-do what is undo here.